F4-Sprint2-TranscriptionSummarization - Move share from tools_repository to workspaces_repository
