+++ controllers/tools_controller.py
@@ -4,7 +4,6 @@ from services.tools_service import ToolsService
 from databases.dependencies import get_session
 from models.requests.transcript_request import TranscriptRequest
 from models.requests.summarize_request import SummarizeRequest
-from models.requests.share_request import ShareRequest
 
 router = APIRouter(prefix = "/api/tools")
 
@@ -17,8 +16,3 @@ async def transcript(request: TranscriptRequest, db: Session = Depends(get_sessi
 async def summarize(request: SummarizeRequest, db: Session = Depends(get_session)):
     tools_service = ToolsService(db)
     return await tools_service.summarize(request)
-
-@router.post("/share")
-async def share(request: ShareRequest, db: Session = Depends(get_session)):
-    tools_service = ToolsService(db)
-    return await tools_service.share(request)
