+++ repositories/tools_repository.py
@@ -3,11 +3,9 @@ from databases.tr_workspace import TrWorkspace
 from databases.tr_workspace_detail import TrWorkspaceDetail
 from models.requests.transcript_request import TranscriptRequest
 from models.requests.summarize_request import SummarizeRequest
-from models.requests.share_request import ShareRequest
 from models.responses.response import Response
 from models.responses.transcript_response import TranscriptResult
 from models.responses.summarize_response import SummarizeResult
-from models.responses.share_response import ShareResult
 from pipelines.summarization_pipeline import model as summarization_model, tokenizer as summarization_tokenizer
 from pipelines.transcription_pipeline import model as transcription_model, processor as transcription_processor
 from pipelines.language_model_pipeline import model as language_model
@@ -15,7 +13,6 @@ from http import HTTPStatus
 from utils.base64_utils import get_safe_base64
 from moviepy.editor import VideoFileClip
 from docx import Document
-from transformers import BatchEncoding
 from dotenv import load_dotenv
 import pdfplumber
 import uuid
@@ -230,41 +227,6 @@ class ToolsRepository:
                     result = summarization
                 )
             )
-        except Exception as e:
-            return Response(
-                statusCode = HTTPStatus.INTERNAL_SERVER_ERROR,
-                message = str(e),
-                payload = None
-            )
-        
-    async def share(self, request: ShareRequest):
-        try:
-            workspace = self.db.exec(
-                select(TrWorkspace).where(TrWorkspace.workspaceID == request.workspaceID)
-            ).first()
-
-            if workspace is None:
-                return Response(
-                    statusCode = HTTPStatus.BAD_REQUEST,
-                    message = str(e),
-                    payload = None 
-                )
-            
-            if request.isGrantAccess:
-                workspace.link = f"{self.client_url}/workspace/{workspace.workspaceID}"
-            else:
-                workspace.link = None
-
-
-            self.db.commit()
-
-            return Response[ShareResult](
-                statusCode = HTTPStatus.CREATED,
-                message = None,
-                payload = ShareResult(
-                    link = workspace.link
-                )
-            ) 
         except Exception as e:
             return Response(
                 statusCode = HTTPStatus.INTERNAL_SERVER_ERROR,
